+++ hnn/vispsd.py
@@ -1,16 +1,23 @@
+import sys 
 import os
-import numpy as np
 from math import sqrt
-from copy import deepcopy
 
-from PyQt5.QtWidgets import QAction, QSizePolicy, QFileDialog
+from PyQt5.QtWidgets import (QSizePolicy, QAction, QFileDialog,
+                             QApplication, qApp)
 from PyQt5.QtGui import QIcon
+
+import numpy as np
+
 import matplotlib.pyplot as plt
 import matplotlib.patches as mpatches
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg
 from matplotlib.figure import Figure
 import matplotlib.gridspec as gridspec
+
 from hnn_core.dipole import average_dipoles, Dipole
+
+from .specfn import MorletSpec
+from .paramrw import get_output_dir
 
 from .DataViewGUI import DataViewGUI
 from .specfn import spec_dpl_kernel, extract_spec
